Changed auth_urls to use names consistent with contrib.auth
In up-stream all URL names are as django.contrib.auth but prefixed
``auth_``. I understand the reason for this, but in Django 1.6
the auth views switch from reversing URLs to, e.g., password reset
complete, using the view name to using the URL name. As an app using
Django Registration, those view names will not exist.

After much thought I decided to remove the auth_  prefix from the URL
names in djang_registration so that it becomes plug-compatible again
with Django.

This means that implementing code will need to change templates that use
the old ``auth_`` prefixed URL names.